requirements.txt
streamlit
langgraph
langchain
langchain-core
langchain-google-genai
pydantic
python-dotenv